requirements.txt
tensorflow>=2.13.0
tensorflow-models>=2.13.0
gradio>=4.19.2
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
labelme>=5.3.0 